Adding some necessary methods to Task class
A method for returning the JAR class loader, so that we can set it on the
NodeClient and NodeServer which will need deserialization. Also added
methods for returning the JAR file path and input path, since we will need
it to send the JAR to slave machines and to read the input using the input
reader respectively.

Also modified static factory method name to "load" from "configure", as
well as some exception checking in it so we return null when we can't
instantiate an instance instead of throwing exceptions.