Adding verbose mode to the MasterApplication
On verbose mode we will print extra information like when an unhandled
message has been received and execution phases of the current task, as
well as reasons for failure, etc. directly on the console. Also adding a
printed notification as well when a slave or client connects or
disconnects to the server (that is done everytime though, not only on
verbose mode).